Rename the console session type to TTY
Although the Linux console is the primary intended use case, the name
of the underlying objects that we interface with is more descriptive.

Additionally, although virtual/pseudo terminals as used by graphical
(X11) terminal emulators are TTYs are indeed unlikely to need locking
(due to preferring to lock the graphical session itself), other uses
of TTY, such as serial console, are valid cases.